style(backend): 修复 main 分支 pre-existing clippy 错误
7 个 lint 错误（lib 编译过但 clippy --all-targets 失败）：

1. auth_handler.rs:672 strip_prefix 改用 .strip_prefix() 替代手写切片
2. init_handler.rs:122 unwrap_or(InitTaskStatus::Failed) 替代闭包
3. inventory_adjustment_service.rs:686 移除无用 .map_err(AppError::from)
4. inventory_adjustment_service.rs:786 vec![] → 数组字面量
5. inventory_stock_service.rs:751 #[allow(clippy::too_many_arguments)] + TODO
6. metrics_service.rs:494 删除 assert!(true)（永远为真）
7. import_export.rs:369-371 assert_eq!(x, true) → assert!(x) / assert!(!x)
8. pagination.rs:30 移除冗余 as u64
9. random.rs:19 移除冗余 as char

死代码项加 #[allow(dead_code)] + TODO：
- services/omni_audit_service.rs:38 secret_key 字段
- utils/cache.rs:49 created_at 字段

删除未使用测试辅助函数：
- services/fixed_asset_service.rs:435 create_test_asset（连带删无用 import）

+++ backend/src/utils/pagination.rs
@@ -27,7 +27,7 @@ where
 
     // 顺序执行：先取当前页数据，再统计总数（避免 Paginator 在并行调用时的借用冲突）
     let items = paginator.fetch_page(page_index).await?;
-    let total = paginator.num_items().await? as u64;
+    let total = paginator.num_items().await?;
 
     Ok((items, total))
 }
